+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added support from Allied Vision Technologies cameras using the Vimba driver.
   In the braid .toml configuration file, specify the camera with `start_backend =
   "vimba"`.
+* Braid can now start saving MKV files in all cameras with a single button.
 * Added support to save raw, uncompressed video to the MKV container format.
 * Save camera gamma to MKV files.
 * For Strand Cam and Braid, simplify defaults so that `cargo build --release` is